tests/conftest.py
import asyncio
import pytest
from httpx import AsyncClient
from app.main import app
from dotenv import load_dotenv
load_dotenv()

@pytest.fixture(scope="session")
def event_loop():
    loop = asyncio.get_event_loop()
    yield loop


@pytest.fixture()
async def client():
    async with AsyncClient(app=app, base_url="http://test") as ac:
        yield ac